Heightmap snippet of Mont Blanc
Image Snippet of Mont Blanc which is greysacled, turned to 16 bit channeld and then "curved"in photoshop to dim out the majority of the white pallete. Saved as RAW file.

After importing RAW file to unity, I used unity terrain tools to smooth out the points (There was a lot of them).